feat: 등기부등본 권리분석 기능 구현 및 고유번호 기반 조회 흐름 통합
- RegisterAnalyzer에 위험 키워드 및 채권최고액 추출 로직 추가
- 갑구·을구 항목 모두 분석 가능하도록 구조 확장
- 고유번호 기반 부동산 등기부 조회 API와 권리분석 연동 완료
- 분석 결과를 RegisterAnalysis 테이블에 저장하고, 최종 응답 DTO에 반영
- HouseBoard pid 기준으로 HouseInfo 주소 조회 및 자동 분석 연계 구현
- 디버깅 로깅 추가: 위험 키워드 감지, 채권최고액 탐지, 보호한도 계산 결과 출력

+++ src/main/java/com/Realty/RealtyWeb/Entity/HouseInfoEntity.java
@@ -5,6 +5,8 @@ import com.Realty.RealtyWeb.enums.*;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "house_info")
 @Getter
@@ -39,10 +41,10 @@ public class HouseInfoEntity {
     private TransactionType transactionType;  // 거래 유형
 
     @Column(nullable = false, columnDefinition = "DECIMAL(15,2)")
-    private Double price;  // 가격
+    private BigDecimal price;  // 가격
 
     @Column(name = "maintenance_fee", nullable = false, columnDefinition = "DECIMAL(10,2)")
-    private Double maintenanceFee;  // 관리비
+    private BigDecimal maintenanceFee;  // 관리비
 
     @Column(nullable = false, length = 255)
     private String address;  // 주소
@@ -51,10 +53,10 @@ public class HouseInfoEntity {
     private String addressDetail;  // 상세 주소
 
     @Column(name = "exclusive_area", nullable = false, columnDefinition = "DECIMAL(10,2)")
-    private Double exclusiveArea;  // 전용 면적 (㎡)
+    private BigDecimal exclusiveArea;  // 전용 면적 (㎡)
 
     @Column(name = "supply_area", nullable = false, columnDefinition = "DECIMAL(10,7)")
-    private Double supplyArea;  // 공급 면적 (㎡)
+    private BigDecimal supplyArea;  // 공급 면적 (㎡)
 
     @Column
     private Integer rooms;  // 방 개수
